goods/requirements.txt
Django
djangorestframework
django_filters
Pillow
drf_yasg